% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/style.R
\name{house_styles}
\alias{house_styles}
\title{House styles registered in this session}
\usage{
house_styles()
}
\value{
A data frame of registered style names and descriptions.
}
\description{
House styles registered in this session
}
\examples{
house_styles()
}
