% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/registry.R
\name{journal_spec}
\alias{journal_spec}
\title{Look up one journal's figure specification}
\usage{
journal_spec(journal)
}
\arguments{
\item{journal}{Registry id, for example \code{"plos_one"}. Use \code{\link[=journals]{journals()}} to
see the available ids.}
}
\value{
An object of class \code{figspec_spec}.
}
\description{
Look up one journal's figure specification
}
\examples{
journal_spec("frontiers")
}
