% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/style.R
\name{remove_house_style}
\alias{remove_house_style}
\title{Remove a registered house style}
\usage{
remove_house_style(name)
}
\arguments{
\item{name}{Name of the style to remove.}
}
\value{
\code{TRUE} invisibly.
}
\description{
Remove a registered house style
}
\examples{
library(ggplot2)
register_house_style("temporary", theme_void())
remove_house_style("temporary")
}
