% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palette.R
\name{figspec_palette}
\alias{figspec_palette}
\title{The colours in a figspec palette}
\usage{
figspec_palette(palette = "okabe_ito", n = NULL)
}
\arguments{
\item{palette}{Palette id, from \code{\link[=figspec_palettes]{figspec_palettes()}}.}

\item{n}{Number of colours to return. Defaults to all of them.}
}
\value{
A character vector of hex colours.
}
\description{
The colours in a figspec palette
}
\examples{
figspec_palette("okabe_ito", 3)
}
