% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palette.R
\name{scale_colour_figspec}
\alias{scale_colour_figspec}
\alias{scale_color_figspec}
\alias{scale_fill_figspec}
\title{Discrete colour and fill scales using a figspec palette}
\usage{
scale_colour_figspec(palette = "okabe_ito", ...)

scale_color_figspec(palette = "okabe_ito", ...)

scale_fill_figspec(palette = "okabe_ito", ...)
}
\arguments{
\item{palette}{Palette id, from \code{\link[=figspec_palettes]{figspec_palettes()}}.}

\item{...}{Passed to \code{\link[ggplot2:discrete_scale]{ggplot2::discrete_scale()}}.}
}
\value{
A ggplot2 scale.
}
\description{
Discrete colour and fill scales using a figspec palette
}
\examples{
library(ggplot2)
ggplot(mtcars, aes(wt, mpg, colour = factor(cyl))) +
  geom_point() +
  scale_colour_figspec()
}
